🔨| error handling on the end points.

+++ backend/controllers/commentController.js
@@ -4,8 +4,11 @@ const validation = require('../validator/commentValidation');
 
 // Create Order Controller
 const createCommentController = async (req, res) => {
+	if (!req.body) return res.status(400).json({ massage: 'the request needs a body.' });
+
 	// create the DB model with inputs
 	const newComment = new Comment(req.body);
+
 	try {
 		// validate input
 		await validation.validateAsync(req.body);
@@ -14,7 +17,7 @@ const createCommentController = async (req, res) => {
 		const savedComment = await newComment.save();
 
 		// set the response
-		res.status(200).json(savedComment);
+		res.status(201).json(savedComment);
 	} catch (err) {
 		// return the err if there is one
 		res.status(500).json(err);
@@ -24,6 +27,8 @@ const createCommentController = async (req, res) => {
 // update comment ( edit comment )
 const updateCommentController = async (req, res) => {
 	try {
+		if (!req.body) return res.status(400).json({ massage: 'the request needs a body.' });
+
 		// validate input
 		await validation.validateAsync(req.body);
 
@@ -47,11 +52,13 @@ const updateCommentController = async (req, res) => {
 // confirm the comment by admin
 const confirmCommentController = async (req, res) => {
 	try {
+		if (!req.params.id) return res.status(400).json({ massage: 'the request needs an Id params.' });
+
 		// find the comment by Id and set the valid property to TRUE.
 		await Comment.findByIdAndUpdate(req.params.id, { $set: { valid: true } }, { new: true });
 
 		// set response
-		res.status(200).json('comment confirmed successfully !');
+		res.status(200).json({ massage: 'comment confirmed successfully !' });
 	} catch (err) {
 		// return the err if there is one
 		res.status(500).json(err);
@@ -61,6 +68,8 @@ const confirmCommentController = async (req, res) => {
 // delete comment
 const deleteCommentController = async (req, res) => {
 	try {
+		if (!req.params.id) return res.status(400).json({ massage: 'the request needs an Id params.' });
+
 		// find the comment by id and delete.
 		const deletedComment = await Comment.findByIdAndDelete(req.params.id);
 
@@ -90,6 +99,8 @@ const getAllUserCommentsController = async (req, res) => {
 // get a comment by its Id
 const getOneCommentByIdController = async (req, res) => {
 	try {
+		if (!req.params.id) return res.status(400).json({ massage: 'the request needs an Id params.' });
+
 		// find the comment By Id
 		const comment = await Comment.findById(req.params.id);
 
@@ -104,6 +115,9 @@ const getOneCommentByIdController = async (req, res) => {
 // get all the comments of one product.
 const getCommentByProductController = async (req, res) => {
 	try {
+		if (!req.params.id)
+			return res.status(400).json({ massage: 'the request needs an product_Id params.' });
+
 		// find the comments of that product By Id
 		const userComments = await Comment.find({ product_Id: req.params.product_Id });
